Better handling of edge cases.
More work on this branch to better handle built-in model downloading. Added a feature to replace the "Generate" button with a "Download" button when the selected built-in model/compute units combination is not yet downloaded. When compute units change the list of models refreshes to a filtered list of models supporting that variant.