Remove Path swift package
Path is useful shorthand for working with URLs. In this project Path is used in 3 files while basic URLs are used in the rest. SwiftUI.Path is a new addition causing occasional compiler issues trying to determine which API is being referenced. This submission removes the Path swift package from the project and updates all instances where Path was used to URL alternative code.