PromptTextField
https://github.com/huggingface/swift-coreml-diffusers/issues/53
Rather than limiting the prompt entry on textfield size limits or character limits this PR adds a token tracker.

As the user enters prompts an indicator displays the current number of used tokens versus the maximum of 75.

Tokens are calculated using the currently selected model.

For the moment this is a short term goal to  expose the token count to the end user with a change of text color to warn of approaching or exceeding mac token count.

Note that in a1111 they handle this by merging excessive tokens. "AUTOMATIC1111 has no token limits. If a prompt contains more than 75 tokens, the limit of the CLIP tokenizer, it will start a new chunk of another 75 tokens, so the new “limit” becomes 150. The process can continue forever or until your computer runs out of memory…

Each chunk of 75 tokens is processed independently, and the resulting representations are concatenated before feeding into Stable Diffusion’s U-Net."