macOS batch image creation
Adding a new batch image creation feature as requested in https://github.com/huggingface/swift-coreml-diffusers/issues/52

note: That there is a missing requested feature.
This PR generates a batch of images.
This PR does NOT yet have the following features:
- select one of the images to share individually (it shares the first generated image)
- define the seed that was used for each image in the batch
- use a starting seed and increment by one for each image in the batch

It DOES serve as a starting point to view the results of multiple images generated in a batch on macOS.