Dragging improvement
Tis PR also includes the ability to drag the image out into other applications on macOS. This commit improves that behavious and catches an issue with how the application support folder is being addressed elsewhere in the project. FileManager should likely be used instead of URL to ensure that the app specific application support folder it used in all instances.