Model Importer
macOS only** (For iOS it would be good to open from Files, Dropbox, and other compatible file sources.)

To trigger this new feature you can open the File menu and select Import Model, you can press Command-Shift-I, or you can press the [+] button next to the list of models.

Once triggered a file browser will show to allow the user to navigate to a folder with a model they downloaded from the Hub / internet.
The user can select a folder or a .zip file. One selected the code will validate whether key necessary files and resources exist inside the selected folder or blindly process a zip file.

If the folder contains the key resources then it copies the folder to the models folder. If it's a zip file that file is extracted into the models folder.

If the folder does not contain the key resources then an error message is displayed. This way we prevent corrupted entries from appearing in the UI. (note: there is a missing check of the extracted zip file at this time.)