Uint32 Seed
Justification: The seed internally is changed between different number types with wildly different max values. This PR normalizes all of them to UInt32.

Earlier @pcuenca commented that the reason for the lower precision use of Double was to support the CompactSlider component and make it easier for a user to select a number.

In this submission the CompactSlider component has been swapped for a TextField.

[Feature extracted from larger PR #59 https://github.com/huggingface/swift-coreml-diffusers/pull/59]