PromptTextField (#58)
* PromptTextField

https://github.com/huggingface/swift-coreml-diffusers/issues/53
Rather than limiting the prompt entry on textfield size limits or character limits this PR adds a token tracker.

As the user enters prompts an indicator displays the current number of used tokens versus the maximum of 75.

Tokens are calculated using the currently selected model.

For the moment this is a short term goal to  expose the token count to the end user with a change of text color to warn of approaching or exceeding mac token count.

Note that in a1111 they handle this by merging excessive tokens. "AUTOMATIC1111 has no token limits. If a prompt contains more than 75 tokens, the limit of the CLIP tokenizer, it will start a new chunk of another 75 tokens, so the new “limit” becomes 150. The process can continue forever or until your computer runs out of memory…

Each chunk of 75 tokens is processed independently, and the resulting representations are concatenated before feeding into Stable Diffusion’s U-Net."

* Update Diffusion-macOS/ControlsView.swift

spacing

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>

* Update Diffusion/Common/Views/PromptTextField.swift

increase the mac token count to account for start and end tokens.

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>

* Moved token count below textfield

On macOS token count now displays below the textfield right aligned. On iOS the token count displays below the text field left aligned. The text color in the text field was changed from .white to .primary to avoid dark mode/light mode issues.

On iOS a bug was fixed where the selected model was not matching due to changes in the way palettized models were being selected.

* Correctly select palettized version when needed

v21Palettized

* Update Diffusion/Common/Views/PromptTextField.swift

Comment spacing style guide compliance.

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>

* Update Diffusion/Common/Views/PromptTextField.swift

Comment spacing style guide compliance.

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>

* Update Diffusion/Common/Views/PromptTextField.swift

update token count on appear to ensure that the model (if present) at load is used for field token calculations.

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>

* Remove duplicated lines in project

---------

Co-authored-by: Pedro Cuenca <pedro@huggingface.co>