Application support path fix (#60)
* Application support path fix

I discovered that post Path.swift fix the location of the default models folder was inconsistent. This PR creates a new internal API to access the user domain applicationSupportDirectory to improve consistency. If there is an issue with the user domain then this code will fall back to URL.applicationSupportDirectory but under normal operating conditions it should use the app bundle.

In the previous code I was using URL.applicationSupportDirectory directly which results in
/Users/<uname>/Library/Application Support/hs-diffusion-models

Whereas the previous and correct usage is
/Users/<uname>/Library/Application Support/<bundle identifier>hf-diffusion-models

I've put the new applicationSupportURL function inside Settings. To access it use Settings.shared.applicationSupportURL()

Is Settings.shared a reasonable location for such an API? It's not quite a Util function.

* Update Diffusion/Common/State.swift

remove bundle identifier in path to ensure backward compatibility with current live release version.